+++ plugins/ibutsu/src/components/IbutsuComponent/IbutsuComponent.tsx
@@ -20,24 +20,18 @@ import Stack from '@mui/material/Stack';
 import { useApi, configApiRef } from '@backstage/core-plugin-api';
 import { useEntity } from '@backstage/plugin-catalog-react';
 import { InfoCard } from '@backstage/core-components';
+import { queryHealthAndRunsData } from '../../common/QueryHealthAndRunsData';
 
 export function IbutsuComponent() {
-    const [results, setResults] = React.useState<any[]>([]);
     const [resultsPage, setResultsPage] = React.useState<any[]>([]);
     const [page, setPage] = React.useState<number>(0);
-    const [infoData, setInfoData] = React.useState<any[]>([]);
-    const [showHealthDataError, setShowHealthDataError] = React.useState(false)
-    const [showRunsDataError, setShowRunsDataError] = React.useState(false)
 
     // state variable for spinner logic
-    const [loading, setLoading] = React.useState(true)
     const pageSize: number = 5
 
-    const config = useApi(configApiRef);
-    const { entity } = useEntity();
     const title = 'Test Results';
 
-    console.log(entity?.metadata?.annotations?.["ibutsu/component-name"]);
+    const { results, infoData, showHealthDataError, showRunsDataError, loading } = queryHealthAndRunsData();
 
     // Style configurations for buttons
     const useStylesButton = makeStyles((theme: Theme) =>
@@ -96,98 +90,23 @@ export function IbutsuComponent() {
     }
 
     const formatISOdateOutput = (isoDateFormat) => {
-        const date = isoDateFormat.substring(0,10)
-        const time = isoDateFormat.substring(11,19)
-    
+        const date = isoDateFormat.substring(0, 10)
+        const time = isoDateFormat.substring(11, 19)
+
         return `${date}, ${time}`
-    }
-
-    const appName = () => {
-        // Some component names don't match the app's name
-        if (entity?.metadata?.annotations?.["ibutsu/component-name"]) {
-            return entity?.metadata?.annotations?.["ibutsu/component-name"]
-        }
-
-        return entity?.metadata?.name.substring(0, entity?.metadata?.name.indexOf("-"));
-    }
-
-    const apiBaseUrl = () => {
-        const baseUrl: string = config.getString('backend.baseUrl');
-        return `${baseUrl}/api/proxy/ibutsu/api`;
-    }
-
-    const getHealthData = async () => {
-        const healthUrl = `${apiBaseUrl()}/health/info`;
-
-        fetch(healthUrl, {
-            method: 'GET',
-            headers: {
-                'Content-Type': 'application/json'
-            },
-        }).then(response => {
-            if (!response.ok) {
-                throw new Error(`Network response was not ok ${response.statusText}`);
-            }
-            return response.json();
-        }).then(data => {
-            setInfoData(data.frontend);
-            console.log(infoData);
-        }).catch(error => {
-            setShowHealthDataError(true);
-            console.error('error fetching health data from Ibutsu API:', error);
-        });
-    }
-
-    const getRunsData = async () => {
-        const runsOpts = {
-            'filter': `component=${appName()}`,
-            'page': 1,
-            'pageSize': 50,
-        };
-
-        const runsQueryString = new URLSearchParams(runsOpts).toString();
-        const runsUrl = `${apiBaseUrl()}/run?${runsQueryString}`;
-
-        fetch(runsUrl, {
-            method: 'GET',
-            headers: {
-                'Content-Type': 'application/json'
-            },
-        }).then(response => {
-            if (!response.ok) {
-                throw new Error(`Network response was not ok ${response.statusText}`);
-            }
-
-            return response.json();
-        }).then(data => {
-            const appRuns = data.runs;
-
-            if (appRuns.length === 0) {
-                throw new Error(
-                    `No test runs data was returned from the Ibutsu API`);
-            }
-
-            setResults(appRuns.filter(d => d.component === appName()))
-            console.log(data);
-        }).catch(error => {
-            setShowRunsDataError(true);
-            console.error('Error fetching test run data from Ibutsu API:', error);
-        }).finally(() => {
-            setLoading(false)
-        });
     }
 
     const RowHead = () => {
         return (
             <TableHead>
-            <TableRow>
-                <TableCell align="center"><Typography variant="button">Created</Typography></TableCell>
-                <TableCell align="center"><Typography variant="button">Run ID</Typography></TableCell>
-                <TableCell align="center"><Typography variant="button">Duration</Typography></TableCell>
-                <TableCell align="center"><Typography variant="button">Environment</Typography></TableCell>
-                <TableCell align="center"><Typography variant="button">Results</Typography></TableCell>
-            </TableRow>
-        </TableHead>
+                <TableRow>
+                    <TableCell align="center"><Typography variant="button">Created</Typography></TableCell>
+                    <TableCell align="center"><Typography variant="button">Run ID</Typography></TableCell>
+                    <TableCell align="center"><Typography variant="button">Duration</Typography></TableCell>
+                    <TableCell align="center"><Typography variant="button">Environment</Typography></TableCell>
+                    <TableCell align="center"><Typography variant="button">Results</Typography></TableCell>
+                </TableRow>
+            </TableHead>
         )
     }
 
@@ -208,7 +127,7 @@ export function IbutsuComponent() {
             {summaryItems.map(item => (
                 result.summary[item.key] > 0 && (
                     <Tooltip title={item.title}>
-                        <Pill style={{backgroundColor: item.color, color: "#FFFFFF"}}>{result.summary[item.key]}</Pill>
+                        <Pill style={{ backgroundColor: item.color, color: "#FFFFFF" }}>{result.summary[item.key]}</Pill>
                     </Tooltip>
                 )
             ))}
@@ -227,7 +146,7 @@ export function IbutsuComponent() {
                 <TableCell align="center">{result.duration.toFixed(1)}s</TableCell>
                 <TableCell align="center">{result.env}</TableCell>
                 <TableCell align="center">
-                    <SummaryPills result={result}/>
+                    <SummaryPills result={result} />
                 </TableCell>
             </TableRow>
         )
@@ -256,10 +175,10 @@ export function IbutsuComponent() {
     }
 
     // fetch data from Ibutsu API on load only
-    useEffect(() => {
-        getHealthData()
-        getRunsData()
-    }, [])
+    //useEffect(() => {
+    //    getHealthData()
+    //    getRunsData()
+    //}, [])
 
     // Used for pagination
     useEffect(() => {
